stream_data/producer_stockdata_fake/requirements.txt
pyspark
pandas
kafka-python
requests
alpaca-py
pytz
nest_asyncio
aiohttp
psycopg2-binary